Refactor Actioncard, board to package session and add House Messaging dtos
