--- src/main/java/at/aau/serg/websocketserver/actioncard/ActionCardController.java
@@ -1,4 +1,0 @@
-package at.aau.serg.websocketserver.actioncard;
-
-public class ActionCardController {
-}
