+++ src/main/java/at/aau/serg/websocketserver/messaging/dtos/HouseBuyElseSellMessage.java
@@ -1,0 +1,18 @@
+package at.aau.serg.websocketserver.messaging.dtos;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class HouseBuyElseSellMessage {
+    private String playerID;
+    private int gameId;
+    private boolean buyElseSell;
+    private int houseId;
+
+
+    public boolean buyElseSell() {
+        return buyElseSell;
+    }
+}
