+++ src/main/java/at/aau/serg/websocketserver/session/actioncard/ActionCardController.java
@@ -1,0 +1,4 @@
+package at.aau.serg.websocketserver.session.actioncard;
+
+public class ActionCardController {
+}
