+++ src/main/java/at/aau/serg/websocketserver/session/board/BoardService.java
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketserver.board;
+package at.aau.serg.websocketserver.session.board;
 
 import org.springframework.stereotype.Service;
