+++ src/main/java/at/aau/serg/websocketserver/session/board/Field.java
@@ -1,4 +1,4 @@
-package at.aau.serg.websocketserver.board;
+package at.aau.serg.websocketserver.session.board;
 
 import java.util.List;
 import java.util.ArrayList;
