+++ src/test/java/at/aau/serg/websocketserver/actioncard/ActionCardDeckTest.java
@@ -1,5 +1,7 @@
 package at.aau.serg.websocketserver.actioncard;
 
+import at.aau.serg.websocketserver.session.actioncard.ActionCard;
+import at.aau.serg.websocketserver.session.actioncard.ActionCardDeck;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
