+++ src/test/java/at/aau/serg/websocketserver/board/BoardServiceTest.java
@@ -1,5 +1,7 @@
 package at.aau.serg.websocketserver.board;
 
+import at.aau.serg.websocketserver.session.board.BoardService;
+import at.aau.serg.websocketserver.session.board.Field;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
