+++ src/test/java/at/aau/serg/websocketserver/board/FieldTest.java
@@ -1,5 +1,6 @@
 package at.aau.serg.websocketserver.board;
 
+import at.aau.serg.websocketserver.session.board.Field;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
